requirements.txt
# Requirements to run the tool
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
pydantic==2.6.4
python-multipart==0.0.9
PyMuPDF==1.23.20
transformers==4.39.3
Pillow==10.3.0
opencv-python-headless==4.9.0.80
requests==2.31.0
torch==2.2.2
torchvision==0.17.2
accelerate==0.28.0
streamlit==1.32.2
easyocr==1.7.1
PyPDF2==3.0.1
matplotlib==3.8.3
scikit-learn==1.4.2
scikit-image==0.22.0
gradio==4.25.0
pytest==8.1.1